+++ libraries/Ticker/src/Ticker.h
@@ -45,13 +45,13 @@ public:
 		_attach_s(seconds, true, _static_callback, this);
 	}
 
-	void attach_ms(uint32_t milliseconds, callback_function_t callback)
+	void attach_ms(uint64_t milliseconds, callback_function_t callback)
 	{
 		_callback_function = std::move(callback);
 		_attach_ms(milliseconds, true, _static_callback, this);
 	}
 
-	void attach_us(uint32_t micros, callback_function_t callback)
+	void attach_us(uint64_t micros, callback_function_t callback)
 	{
 		_callback_function = std::move(callback);
 		_attach_us(micros, true, _static_callback, this);
@@ -68,14 +68,14 @@ public:
 	}
 
 	template<typename TArg>
-	void attach_ms(uint32_t milliseconds, void (*callback)(TArg), TArg arg)
+	void attach_ms(uint64_t milliseconds, void (*callback)(TArg), TArg arg)
 	{
 		static_assert(sizeof(TArg) <= sizeof(void*), "attach() callback argument size must be <= sizeof(void*)");
 		_attach_ms(milliseconds, true, reinterpret_cast<callback_with_arg_t>(callback), (void*)arg);
 	}
 
 	template<typename TArg>
-	void attach_us(uint32_t micros, void (*callback)(TArg), TArg arg)
+	void attach_us(uint64_t micros, void (*callback)(TArg), TArg arg)
 	{
 		static_assert(sizeof(TArg) <= sizeof(void*), "attach() callback argument size must be <= sizeof(void*)");
 		_attach_us(micros, true, reinterpret_cast<callback_with_arg_t>(callback), (void*)arg);
@@ -87,13 +87,13 @@ public:
 		_attach_s(seconds, false, _static_callback, this);
 	}
 
-	void once_ms(uint32_t milliseconds, callback_function_t callback)
+	void once_ms(uint64_t milliseconds, callback_function_t callback)
 	{
 		_callback_function = std::move(callback);
 		_attach_ms(milliseconds, false, _static_callback, this);
 	}
 
-	void once_us(uint32_t micros, callback_function_t callback)
+	void once_us(uint64_t micros, callback_function_t callback)
 	{
 		_callback_function = std::move(callback);
 		_attach_us(micros, false, _static_callback, this);
@@ -107,14 +107,14 @@ public:
 	}
 
 	template<typename TArg>
-	void once_ms(uint32_t milliseconds, void (*callback)(TArg), TArg arg)
+	void once_ms(uint64_t milliseconds, void (*callback)(TArg), TArg arg)
 	{
 		static_assert(sizeof(TArg) <= sizeof(void*), "attach() callback argument size must be <= sizeof(void*)");
 		_attach_ms(milliseconds, false, reinterpret_cast<callback_with_arg_t>(callback), (void*)arg);
 	}
 
 	template<typename TArg>
-	void once_us(uint32_t micros, void (*callback)(TArg), TArg arg)
+	void once_us(uint64_t micros, void (*callback)(TArg), TArg arg)
 	{
 		static_assert(sizeof(TArg) <= sizeof(void*), "attach() callback argument size must be <= sizeof(void*)");
 		_attach_us(micros, false, reinterpret_cast<callback_with_arg_t>(callback), (void*)arg);
@@ -124,7 +124,7 @@ public:
 	bool active() const;
 
 protected:
-	void _attach_ms(uint32_t milliseconds, bool repeat, callback_with_arg_t callback, void* arg);
+	void _attach_ms(uint64_t milliseconds, bool repeat, callback_with_arg_t callback, void* arg);
 	static void _static_callback(void* arg);
 
 	callback_function_t _callback_function = nullptr;
@@ -132,7 +132,7 @@ protected:
 	esp_timer_handle_t _timer;
 
 private:
-	void _attach_us(uint32_t micros, bool repeat, callback_with_arg_t callback, void* arg);
+	void _attach_us(uint64_t micros, bool repeat, callback_with_arg_t callback, void* arg);
 	void _attach_s(float seconds, bool repeat, callback_with_arg_t callback, void* arg);
 };
